Add permissions to dependent endpoints

+++ app/Http/Controllers/UserController.php
@@ -13,13 +13,13 @@ use DB;
 
 class UserController extends Controller
 {
-     public function __construct() 
+     public function __construct()
      {
          // $this->middleware('jwt.auth', ['except' => ['authenticate', 'token']]);
         // provides an authorization header with each response
         // $this->middleware('jwt.refresh', ['except' => ['authenticate', 'token']]);
      }
-    
+
     /*
      * Get a validator for an incoming registration request.
      */
@@ -70,12 +70,15 @@ class UserController extends Controller
     // TODO: isOwner() and isAdmin(), to set permission/role filters on managing dependents
     //
 
-    /* 
+    /*
      * GET api/accounts/{id}/dependents
      * - list all dependents
      */
      public function index($accountID)
     {
+        if (!Entrust::can(PermissionNames::ApproveUserRegistration()) && Auth::user()->id != $accountID) {
+            return response()->json(["message" => "no_user_approval_access"]);
+        }
         // TODO: add permission/role filter
         $account = Account::where('id', '=', $accountID)->first();
         if ($account === null)
@@ -94,9 +97,12 @@ class UserController extends Controller
      */
     public function addDependents($accountID, Request $req)
     {
+        if (!Entrust::can(PermissionNames::ApproveUserRegistration()) && Auth::user()->id != $accountID) {
+            return response()->json(["message" => "no_user_approval_access"]);
+        }
         $dependents = $req->all();
         try
-        {   
+        {
             DB::beginTransaction();
             foreach ($dependents as $d)
             {
@@ -123,6 +129,9 @@ class UserController extends Controller
      */
     public function editDependent($accountID, $depId, Request $req)
     {
+        if (!Entrust::can(PermissionNames::ApproveUserRegistration()) && Auth::user()->id != $accountID) {
+            return response()->json(["message" => "no_user_approval_access"]);
+        }
         $changes = $req->all();
         // dd($changes);
         $user = User::where('id', $depId)
@@ -130,7 +139,7 @@ class UserController extends Controller
                     ->first();
         if (!$user)
             return response()->json(['message' => 'user_does_not_exist'], 422);
-        else 
+        else
         {
             $validator = $this->dependentEditValidator($changes);
             if ($validator->passes())
@@ -157,6 +166,7 @@ class UserController extends Controller
                 return response()->json(['message' => 'validation_failed', 'errors' => $validator->errors()], 422);
         }
 
+        $user->approved = false;
         if ($user->save())
             return response()->json(['message' => 'user_updated'], 200);
         else
@@ -169,6 +179,10 @@ class UserController extends Controller
      */
     public function deleteDependent($accountID, $depId)
     {
+        if (!Entrust::can(PermissionNames::ApproveUserRegistration()) && Auth::user()->id != $accountID) {
+            return response()->json(["message" => "no_user_approval_access"]);
+        }
+
         $user = User::where('accountID', '=', $accountID)
                     ->where('id', '=', $depId);
         if ($user->delete())
